feat code cleanup
