+++ src/Constant/Application/Common.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 export const Countries = [
   { label: 'Japan', value: 116 },
   { label: 'Australia', value: 12 },
@@ -28,163 +27,3 @@ export const cargoSize = [
   { label: '1.125', value: '1.125' },
   { label: '1.250', value: '1.250' },
 ];
-
-export const TextInputTypes = {
-  DEFAULT: 'default',
-  BOX_NO: 'boxno',
-  CONTACT_PH: 'contact_ph',
-};
-
-export const RegexExpression = {
-  default: '([A-Z])',
-  boxno: '^([0-9]{3}[A-Z]{2}[0-9]{3})',
-  contact_ph: `^(09)\\d{9}$`,
-};
-
-export const InvalidFeedback = {
-  boxno: 'Incorrect Format, Example: 123AB123',
-  contact_ph: 'Incorrect Format, Example: 09396341125',
-};
-
-export const Module = {
-  PACKINGLIST: 'packinglist',
-  EMPTYBOX: 'emptybox',
-  PICKUPS: '',
-  DISPATCH: 'dispatch',
-  FLEET: 'fleet',
-};
-
-export const moduleInitialState = {
-  packinglist: {
-    origin: 116,
-    cargoType: 1,
-    company: null,
-    agent: null,
-    boxNumber: null,
-    trackingNumber: null,
-    forwarder: null,
-    size: null,
-    equivalent: null,
-    lengthWidthHeight: null,
-    cbm: null,
-    serialNumber: null,
-    lastName_s: null,
-    firstName_s: null,
-    contactNumber_s: null,
-    postalCode_s: null,
-    prefecture: null,
-    city_jp: null,
-    town_jp: null,
-    adderss_s: null,
-    lastName_c: null,
-    firstName_c: null,
-    contactNumber_c: null,
-    postalCode_c: null,
-    province: null,
-    city_ph: null,
-    address_c: null,
-    memo: null,
-    deliveryNotes: null,
-  },
-};
-
-export const tableColumnConfig = {
-  packinglist: [
-    {
-      accessorKey: 'origin',
-      header: 'Origin',
-      maxSize: 180,
-      minSize: 180,
-    },
-    {
-      accessorKey: 'company',
-      header: 'Company',
-      maxSize: 190,
-      minSize: 190,
-    },
-    {
-      accessorKey: 'agent',
-      header: 'Agent',
-      maxSize: 190,
-      minSize: 190,
-    },
-    {
-      accessorKey: 'boxNumber',
-      header: 'Box #',
-      maxSize: 150,
-      minSize: 150,
-    },
-    {
-      accessorKey: 'size',
-      header: 'Size',
-      maxSize: 100,
-      minSize: 100,
-    },
-    {
-      accessorKey: 'equivalent',
-      header: 'Equivalent',
-      maxSize: 160,
-      minSize: 160,
-    },
-    {
-      accessorFn: (row: any) =>
-        `${row.lastName_s}, ${row.firstName_s} | ${row.contactNumber_s} | ${row.prefecture}, ${row.city_jp}, ${row.town_jp}, ${row.adderss_s}`,
-      header: 'Sender',
-      maxSize: 550,
-      minSize: 550,
-    },
-  ],
-  emptybox: [],
-  dispatch: [
-    {
-      accessorKey: 'dispatchId',
-      header: 'Dispatch #',
-    },
-    {
-      accessorKey: 'isWhseToPier',
-      header: 'Warehouse > Pier',
-    },
-    {
-      accessorKey: 'boxCount',
-      header: 'Boxes Carried',
-    },
-    {
-      accessorFn: (row: any) => `${row.startDate} | ${row.endDate}`,
-      header: 'Delivery',
-    },
-    {
-      accessorKey: 'fleet',
-      header: 'Fleet #',
-    },
-    {
-      accessorKey: 'truckingCompany',
-      header: 'Trucking Company',
-    },
-    {
-      accessorKey: 'status',
-      header: 'Status',
-    },
-  ],
-  fleet: [
-    {
-      accessorKey: 'truckingCompany',
-      header: 'Company',
-    },
-    {
-      accessorKey: 'plateNumber',
-      header: 'Plate #',
-    },
-    {
-      accessorFn: (row: any) => `${row.driver} | ${row.driverContactNumber}`,
-      header: 'Driver',
-    },
-    {
-      accessorFn: (row: any) => `${row.helper} | ${row.helperContactNumber}`,
-      header: 'Helper',
-    },
-    {
-      accessorKey: 'status',
-      header: 'Status',
-    },
-  ],
-};
