+++ src/Constant/Application/Http.ts
@@ -1,0 +1,6 @@
+export const MethodTypes = {
+  GET: 'GET',
+  POST: 'POST',
+  PUT: 'PUT',
+  PATCH: 'PATCH',
+};
