+++ src/Constant/Application/Module.ts
@@ -1,0 +1,41 @@
+export const Module = {
+  PACKINGLIST: 'packinglist',
+  EMPTYBOX: 'emptybox',
+  PICKUPS: '',
+  DISPATCH: 'dispatch',
+  FLEET: 'fleet',
+};
+
+export const moduleInitialState = {
+  packinglist: {
+    origin: 116,
+    cargoType: 1,
+    company: null,
+    agent: null,
+    boxNumber: null,
+    trackingNumber: null,
+    forwarder: null,
+    size: null,
+    equivalent: null,
+    lengthWidthHeight: null,
+    cbm: null,
+    serialNumber: null,
+    lastName_s: null,
+    firstName_s: null,
+    contactNumber_s: null,
+    postalCode_s: null,
+    prefecture: null,
+    city_jp: null,
+    town_jp: null,
+    adderss_s: null,
+    lastName_c: null,
+    firstName_c: null,
+    contactNumber_c: null,
+    postalCode_c: null,
+    province: null,
+    city_ph: null,
+    address_c: null,
+    memo: null,
+    deliveryNotes: null,
+  },
+};
